Rename PreferenceList.js to confrom to conventions.
