+++ client/lib/preferences-helper/index.js
@@ -9,7 +9,7 @@ import React from 'react';
 /**
  * Internal dependencies
  */
-import PreferenceList from './PreferenceList';
+import PreferenceList from './preference-list';
 
 export default function injectPreferenceHelper( element, store ) {
 	ReactDom.render( React.createElement( PreferenceList, { store } ), element );
